refactor: rename UserEntity and UserRepository for improved package structure
